improve solar eclipse catalog validation
